migration/sqlite-to-postgres/requirements.txt
psycopg2-binary>=2.9.0
dotenv
tqdm>=4.65.0
